⚡️ 지역별 발전량 api 연결 코드 보완
